Added updating your status II
This so needs to be a new topic under Communication but I don't know how to do that yet. There are a few places that need a link to them from How we do Leave but I'll do those once this is in its correct place. (I've corrected the extra space - absolutely no idea if this second attempt is going to work!!)